ReviewBot: provide staging_api() from repo_checker.
